created separate function for addPOI validator
