+++ server/controllers/poiControllers.js
@@ -27,89 +27,39 @@ const { getImageLinks, removeImage } = require("../utils/images");
  *    heading {float}
  *    speed {float}
  */
-exports.addPOI = [
-  // Points of Interest validation
-  body().custom((value, { req }) => {
-    // Add the files back in to the req.body so that they can be treated normally in validation
-    const files = req.files.image;
-    if (files) {
-      value.files = files[0];
-    }
-    return true;
-  }),
-  // required fields
-  body("trailId").exists().isInt().toInt(),
-  body("description", "Invalid data type, must be a string")
-    .exists()
-    .isString()
-    .trim()
-    .escape(),
-  body("files")
-    .exists()
-    .custom((value) => {
-      // check valid mime types
-      const mimetypeArr = value.mimetype.split("/");
+exports.addPOI = async (req, res, next) => {
+  // console.log("req.body", req.body);
+  const controller = "addPOI";
 
-      return (
-        mimetypeArr[0] === "image" &&
-        VALID_IMAGE_TYPES.indexOf(mimetypeArr[1]) > -1
-      );
-    })
-    .withMessage(`Files must be of type: .${VALID_IMAGE_TYPES.join(", .")}`),
-  body("isActive", "Point of Interest isActive must be true/false")
-    .exists()
-    .isBoolean()
-    .toBoolean(),
-  body(["latitude", "longitude"], "Invalid data type, must be a float (#.#)")
-    .exists()
-    .isFloat()
-    .toFloat(),
-  // optional fields
-  body(
-    ["accuracy", "altitude", "altitudeAccuracy", "heading", "speed"],
-    "Invalid data type, must be a float (#.#)"
-  )
-    .optional()
-    .isFloat()
-    .toFloat(),
-
-  // main function
-  async (req, res) => {
-    // console.log("req.body", req.body);
-    const controller = "addPOI";
-
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        logger.error(errors.array(), {
-          controller,
-          errorMsg: "validation error",
-        });
-        res.status(400).json({ error: validationErrors(errors.array()) });
-        return;
-      }
-      // make POI to save
-      const body = req.body;
-      body.image = await getImageLinks(body.trailId, body.files, "POI");
-
-      // save POI to DB
-      const poi = await PointsOfInterest.create(body);
-      if (!poi) {
-        throw new Error("No points of interest created.");
-      }
-
-      res.status(201).json({ poi });
-      return;
-    } catch (err) {
-      logger.error(err, {
+  try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      logger.error(errors.array(), {
         controller,
-        errorMsg: "catch error",
+        errorMsg: "validation error",
       });
-      res.status(500).json({ error: err.message });
+      res.status(400).json({ error: validationErrors(errors.array()) });
       return;
     }
-  },
-];
+    // make POI to save
+    const body = req.body;
+    body.image = await getImageLinks(body.trailId, body.files, "POI");
+
+    // save POI to DB
+    const poi = await PointsOfInterest.create(body);
+    if (!poi) {
+      throw new Error("No points of interest created.");
+    }
+
+    next();
+  } catch (err) {
+    logger.error(err, {
+      controller,
+      errorMsg: "catch error",
+    });
+    res.status(500).json({ error: err.message });
+  }
+};
 
 /**
  * Change an existing point of interest in the DB
